requirements.txt
streamlit
numpy
pandas
snowflake-connector-python